Modularize builder package: extract `buildRemoteIO`, `buildPipeline`, `buildTaskEnqueuer`, and `buildWorkflow` into dedicated files for improved separation of concerns and maintainability.
